fix imports from renaming
